remove base stub module
